rework the license validator for easier cross-project sharing
